Implmented Docker Attach (API, Port REST server)
Implemented docker attach on the Docker API server and Port Layer's
REST server.  Attach involved,

1. Docker API Inspect, Resize, Attach
2. Portlayer hijack middleware (bypasses Swagger for attach)
3. Portlayer hookup for the docker API equivalent above

Also, Sync'd the VIC backend code to match Docker's v1.11 sources
that was recently vendored.  They were mostly function signatures.

[WIP] "3 streams" attach implementation

This is a proof of concept for using Swagger to implement 3
individual streams for stdio from the container.  Instead of
using a middleware and hijacking in the portlayer, it tries
to play nice with standard swagger mechanisms.